requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
cache==1.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
packaging==24.1
pydantic==2.8.2
pymongo==4.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
urllib3==2.2.1
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zipp==3.19.0

colorlog~=6.8.2